+++ apis/v1beta1/collector_webhook.go
@@ -79,13 +79,8 @@ func (c CollectorWebhook) Default(_ context.Context, obj runtime.Object) error {
 		otelcol.Spec.TargetAllocator.Replicas = &one
 	}
 
-	if otelcol.Spec.TargetAllocator.Enabled {
-		TAUnifyEnvVarExpansion(otelcol)
-	}
-
-	if err := DefaultOTLPAddress(otelcol); err != nil {
-		return err
-	}
+	TAUnifyEnvVarExpansion(otelcol)
+	ComponentUseLocalHostAsDefaultHost(otelcol)
 
 	if otelcol.Spec.Autoscaler != nil && otelcol.Spec.Autoscaler.MaxReplicas != nil {
 		if otelcol.Spec.Autoscaler.MinReplicas == nil {
@@ -476,64 +471,35 @@ func TAUnifyEnvVarExpansion(otelcol *OpenTelemetryCollector) {
 
 	const (
 		baseFlag = "feature-gates"
-		fgFlag   = "-confmap.unifyEnvVarExpansion"
+		fgFlag   = "confmap.unifyEnvVarExpansion"
 	)
 	if otelcol.Spec.Args == nil {
 		otelcol.Spec.Args = make(map[string]string)
 	}
 	args, ok := otelcol.Spec.Args[baseFlag]
 	if !ok || len(args) == 0 {
-		otelcol.Spec.Args[baseFlag] = fgFlag
+		otelcol.Spec.Args[baseFlag] = "-" + fgFlag
 	} else if !strings.Contains(otelcol.Spec.Args[baseFlag], fgFlag) {
-		otelcol.Spec.Args[baseFlag] += "," + fgFlag
+		otelcol.Spec.Args[baseFlag] += ",-" + fgFlag
 	}
-
 }
 
-// DefaultOTLPAddress binds configured otlp receivers to 0.0.0.0 if nothing else
-// is explicitly defined.
-func DefaultOTLPAddress(otelcol *OpenTelemetryCollector) error {
-	for key, rc := range otelcol.Spec.Config.Receivers.Object {
-		// check if otel is configured
-		if !strings.HasPrefix(key, "otlp") {
-			continue
-		}
-
-		cfg, ok := rc.(map[string]interface{})
-		if !ok {
-			continue
-		}
-
-		protocols, ok := cfg["protocols"].(map[string]interface{})
-		if !ok {
-			continue
-		}
-
-		if _, exists := protocols["grpc"]; exists {
-			grpcConfig, ok := protocols["grpc"].(map[string]interface{})
-			if !ok {
-				grpcConfig = make(map[string]interface{})
-				protocols["grpc"] = grpcConfig
-			}
-
-			gEndpoint, ok := grpcConfig["endpoint"].(string)
-			if !ok || !strings.Contains(gEndpoint, ":") {
-				grpcConfig["endpoint"] = "0.0.0.0:4317"
-			}
-		}
-
-		if _, exists := protocols["http"]; exists {
-			httpConfig, ok := protocols["http"].(map[string]interface{})
-			if !ok {
-				httpConfig = make(map[string]interface{})
-				protocols["http"] = httpConfig
-			}
-
-			hEndpoint, ok := httpConfig["endpoint"].(string)
-			if !ok || !strings.Contains(hEndpoint, ":") {
-				httpConfig["endpoint"] = "0.0.0.0:4318"
-			}
-		}
+// ComponentUseLocalHostAsDefaultHost enables component.UseLocalHostAsDefaultHost
+// featuregate on the given collector instance.
+// NOTE: For more details, visit:
+// https://github.com/open-telemetry/opentelemetry-collector/issues/8510
+func ComponentUseLocalHostAsDefaultHost(otelcol *OpenTelemetryCollector) {
+	const (
+		baseFlag = "feature-gates"
+		fgFlag   = "component.UseLocalHostAsDefaultHost"
+	)
+	if otelcol.Spec.Args == nil {
+		otelcol.Spec.Args = make(map[string]string)
 	}
-	return nil
+	args, ok := otelcol.Spec.Args[baseFlag]
+	if !ok || len(args) == 0 {
+		otelcol.Spec.Args[baseFlag] = "-" + fgFlag
+	} else if !strings.Contains(otelcol.Spec.Args[baseFlag], fgFlag) {
+		otelcol.Spec.Args[baseFlag] += ",-" + fgFlag
+	}
 }